Refactor sd
